refactor(components): refactor collapse / collapse-item
