--- packages/components/collapse/src/collapse.type.ts
@@ -1,6 +1,0 @@
-import type { Ref } from 'vue'
-
-export interface CollapseProvider {
-  activeNames: Ref
-  handleItemClick: (name: string | number) => void
-}
